Partly rewrite Redux actions
