+++ eisbuk-admin/src/store/actions/actions.tsx
@@ -8,48 +8,13 @@ import { Slot, Customer } from "eisbuk-shared";
 
 import { Action, NotifVariant } from "@/enums/Redux";
 
-import {
-  CustomerInStore,
-  Dispatch,
-  GetState,
-  Notification,
-  SlotDay,
-  SlotWeek,
-} from "@/types/store";
+import { Dispatch, GetState } from "@/types/store";
 
 import { functionsZone, ORGANIZATION } from "@/config/envInfo";
 
 interface GetFirebase {
   getFirebase: () => typeof firebase;
 }
-
-const enqueueSnackbar = (notification: Notification) => {
-  const key = notification.key || new Date().getTime() + Math.random();
-
-  return {
-    type: Action.EnqueueSnackbar,
-    notification: {
-      ...notification,
-      key,
-    },
-  };
-};
-
-const closeSnackbar = (key: Notification["key"]) => ({
-  type: Action.CloseSnackbar,
-  dismissAll: !key, // dismiss all if no key has been defined
-  key,
-});
-
-export const removeSnackbar = (
-  key: Notification["key"]
-): {
-  type: Action;
-  key: Notification["key"];
-} => ({
-  type: Action.RemoveSnackbar,
-  key,
-});
 
 export const signOut = () => async (
   dispatch: Dispatch,
@@ -377,131 +342,3 @@ export const markAbsentee = ({
     .doc(slot.id)
     .set({ absentees: { [user.id]: isAbsent } }, { merge: true });
 };
-
-export const deleteCustomer = (customer: CustomerInStore) => async (
-  dispatch: Dispatch,
-  getState: GetState,
-  { getFirebase }: GetFirebase
-): Promise<void> => {
-  const firebase = getFirebase();
-
-  try {
-    await firebase
-      .firestore()
-      .collection("organizations")
-      .doc(ORGANIZATION)
-      .collection("customers")
-      .doc(customer.id)
-      .delete();
-
-    dispatch(
-      enqueueSnackbar({
-        key: new Date().getTime() + Math.random(),
-        message: `${customer.name} ${customer.surname} rimosso`,
-        options: {
-          variant: NotifVariant.Success,
-          action: (key) => (
-            <Button
-              variant="outlined"
-              onClick={() => dispatch(closeSnackbar(key))}
-            >
-              OK
-            </Button>
-          ),
-        },
-      })
-    );
-  } catch {
-    showErrSnackbar(dispatch);
-  }
-};
-
-const showErrSnackbar = (dispatch: Dispatch): void => {
-  dispatch(
-    enqueueSnackbar({
-      message: "Errore",
-      options: {
-        variant: NotifVariant.Error,
-      },
-    })
-  );
-};
-
-export const updateCustomer = (customer: CustomerInStore) => async (
-  dispatch: Dispatch,
-  getState: GetState,
-  { getFirebase }: GetFirebase
-): Promise<void> => {
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const { id, ...updatedData } = { ...customer };
-
-  const firebase = getFirebase();
-
-  try {
-    await firebase
-      .firestore()
-      .collection("organizations")
-      .doc(ORGANIZATION)
-      .collection("customers")
-      .doc(customer.id || undefined)
-      .set(updatedData);
-    dispatch(
-      enqueueSnackbar({
-        key: new Date().getTime() + Math.random(),
-        message: `${customer.name} ${customer.surname} aggiornato`,
-        options: {
-          action: (key) => (
-            <Button
-              variant="outlined"
-              onClick={() => dispatch(closeSnackbar(key))}
-            >
-              OK
-            </Button>
-          ),
-        },
-      })
-    );
-  } catch {
-    showErrSnackbar(dispatch);
-  }
-};
-
-export const copySlotDay = (
-  slotDay: SlotDay
-): {
-  type: Action;
-  payload: SlotDay;
-} => ({
-  type: Action.CopySlotDay,
-  payload: slotDay,
-});
-
-export const copySlotWeek = (
-  slotWeek: SlotWeek
-): {
-  type: Action;
-  payload: SlotWeek;
-} => ({
-  type: Action.CopySlotWeek,
-  payload: slotWeek,
-});
-
-export const deleteSlotFromClipboard = (
-  id: Slot<"id">["id"]
-): {
-  type: Action;
-  payload: Slot<"id">["id"];
-} => ({
-  type: Action.DeleteSlotFromClipboard,
-  payload: id,
-});
-
-export const addSlotToClipboard = (
-  slot: Slot<"id">
-): {
-  type: Action;
-  payload: Slot<"id">;
-} => ({
-  type: Action.AddSlotToClipboard,
-  payload: slot,
-});
