+++ eisbuk-admin/src/store/reducers/appReducer.ts
@@ -48,7 +48,7 @@ const appReducer = (
       return {
         ...state,
         notifications: state.notifications.map((notification) =>
-          action.dismissAll || notification.key === action.key
+          !action.key || notification.key === action.key
             ? { ...notification, dismissed: true }
             : { ...notification }
         ),
