+++ eisbuk-admin/src/utils/Notifier.tsx
@@ -4,7 +4,7 @@ import { useSnackbar } from "notistack";
 
 import { LocalStore, Notification } from "@/types/store";
 
-import { removeSnackbar } from "@/store/actions/actions";
+import { removeSnackbar } from "@/store/actions/notifications";
 
 /** @TODO refactor to use imported selector */
 const selectNotifications = (store: LocalStore) =>
